+++ gateway/src/apicast/oauth/oidc.lua
@@ -105,7 +105,7 @@ end
 
 local function find_jwk(jwt, keys)
   local jwk = keys and keys[jwt.header.kid]
-  if jwk then return jwk end
+  return jwk
 end
 
 -- Parses the token - in this case we assume it's a JWT token
@@ -185,8 +185,13 @@ function _M:verify(jwt, cache_key)
   -- Find jwk with matching kid for current JWT in request
   local jwk_obj = find_jwk(jwt, self.keys)
 
+  if jwk_obj == nil then
+    ngx.log(ngx.ERR, "[jwt] failed verification for kid: ", jwt.header.kid)
+    return false, '[jwk] not found, token might belong to a different realm'
+  end
+
   local pubkey = jwk_obj.pem
-  -- Check the jwk for the alg field and if not present skip the validation as it is 
+  -- Check the jwk for the alg field and if not present skip the validation as it is
   -- OPTIONAL according to https://www.rfc-editor.org/rfc/rfc7517#section-4.4
   if jwk_obj.alg and jwk_obj.alg ~= jwt.header.alg then
     return false, '[jwt] alg mismatch'